Completely remove the Rails application
There's nothing in Movie Terminals that requires a backend, even one as
nice as Rails. Initially I considered parsing each terminal command
through Ruby objects to generate the responses. That approach would have
many merits:

- much easier creation of new terminal logic
- easy testing of the command/response logic
- easy logging of failed commands as ideas to add

Ultimately I decided on sticking with the fast response and pure
JavaScript solution of using the CLI library and framework used by
uni.xkcd.com to great effect.

This means that the terminal responses are fast, the setup of making a
"terminal" is already done (although it isn't mobile friendly), and all
logic is entirely contained in client-side JS.

It's the last point that made me finally commit to switching to a static
site generator instead of hosting an actual web application. Movie
Terminals is a toy site: I don't want to serve it up in an application
that I'm not going to maintain. For instance: it's been offline for
almost two years because I was slow in upgrading away from a vulnerable
Rails 3.0 application and Heroku politely asked that it be shut down.

That's another thing: I don't ever intend to pay money to host Movie
Terminals. Keeping all the logic entirely in HTML/JS/CSS means that I
can put it up on my shared Dreamhost space and still easily handle
thousands of vistors. Serving files is easy!

tl;dr: bye Rails, we're gonna switch to a static site