Initial Commit: Jekyll version
There we go! Now Movie Terminals is just a static site generated by
Jekyll. All the stuff that used to be instance variables from the Rails
app for the views is now front YAML matter parsed by Jekyll. The layout
structure is essentially the same. The asset bundle layouts from Rails
are even still essentially the same thanks to the jekyll-assets gem!

There's still a lot of work to be done, but now that the site in back
online and with a more sensible backend work will be a bit more
straightforward.

This is currently the original Movie Terminals site as it was in Rails:
just in Jekyll.

Near term todo:

- less antique and bootstrapy styling for the non-terminal pages
- script to generate a new terminal to make contribution easier
- log failed commands with Clicky Analytics (if possible, I had that
  setup for Google Analytics and it worked great)
- update README on Github with contribution info

Research:

- it'd be awesome to bring in something to replace the JavaScript
  testing setup I had in the Rails version